expint-package {expint}	R Documentation

Exponential Integral and Incomplete Gamma Function

Description

The exponential integrals E_1(x), E_2(x), E_n(x) and Ei(x), and the incomplete gamma function G(a, x) that is defined for negative values of its first argument.

Details

The exponential integral

E_1(x) = int_x^Inf exp(-t)/t dt

and the incomplete gamma function

G(a, x) = int_x^Inf t^(a-1) exp(-t) dt

are closely related functions that arise in various fields of mathematics.

expint is a small package that provides R functions to compute the exponential integral and the incomplete gamma function.

Most conveniently for R package developers, the package also gives access to the underlying C workhorses through an API; see the package vignette for instructions.

The C routines are adapted versions of those of the GNU Scientific Library https://www.gnu.org/software/gsl/.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca

References

Abramowitz, M. and Stegun, I. A. (1972), Handbook of Mathematical Functions, Dover.

See Also

expint for the exponential integral family of functions.

gammainc for the incomplete gamma function.

vignette("expint") for a detailed presentation of the package.

[Package expint version 0.1-6 Index]